backend-evolution/requirements.txt
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.34.0,<1.0
redis>=5.0.0,<6.0
supabase>=2.28.0,<3.0
openai>=1.50.0,<2.0
httpx>=0.28.0,<1.0
python-dotenv>=1.0.0,<2.0
python-multipart>=0.0.20,<1.0
pydantic-settings>=2.5.0,<3.0
